Projects/SlideGeneratorAPI/pyproject.toml
[tool.poetry]
name = "presentation-api"
version = "0.1.0"
description = "AI-powered presentation generation API"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
celery = "^5.3.4"
redis = "^5.0.0"
httpx = "^0.25.2"
openai = "^0.28.1"
python-pptx = "^0.6.23"
cachetools = "^5.3.2"
fastapi-limiter = "^0.1.5"
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.3"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"